+++ payments-core-testing/src/main/java/com/stripe/android/testing/AbsFakeStripeRepository.kt
@@ -27,6 +27,7 @@ import com.stripe.android.model.PaymentMethodCreateParams
 import com.stripe.android.model.PaymentMethodMessage
 import com.stripe.android.model.PaymentMethodUpdateParams
 import com.stripe.android.model.RadarSession
+import com.stripe.android.model.RadarSessionWithHCaptcha
 import com.stripe.android.model.SetupIntent
 import com.stripe.android.model.ShippingInformation
 import com.stripe.android.model.Source
@@ -275,7 +276,7 @@ abstract class AbsFakeStripeRepository : StripeRepository {
 
     override suspend fun createRadarSession(
         requestOptions: ApiRequest.Options
-    ): Result<RadarSession> {
+    ): Result<RadarSessionWithHCaptcha> {
         TODO("Not yet implemented")
     }
 
@@ -284,7 +285,7 @@ abstract class AbsFakeStripeRepository : StripeRepository {
         hcaptchaToken: String,
         hcaptchaEKey: String?,
         requestOptions: ApiRequest.Options
-    ): Result<RadarSession> {
+    ): Result<RadarSessionWithHCaptcha> {
         TODO("Not yet implemented")
     }