+++ payments-core/src/main/java/com/stripe/android/StripeKtx.kt
@@ -520,6 +520,8 @@ suspend fun Stripe.createRadarSession(
                     stripeAccount = stripeAccountId
                 )
             )
+        }.map { radarSession ->
+            RadarSession(id = radarSession.id)
         }
     }
 }