+++ payments-core/src/main/java/com/stripe/android/model/RadarSession.kt
@@ -5,7 +5,5 @@ import kotlinx.parcelize.Parcelize
 
 @Parcelize
 data class RadarSession(
-    val id: String,
-    val passiveCaptchaSiteKey: String?,
-    val passiveCaptchaRqdata: String?
+    val id: String
 ) : StripeModel